+++ Editor/CustomInspectors/BaseSpriteBundleLoaderEditor.cs
@@ -9,8 +9,7 @@ using Utils;
 
 namespace BundlesLoader.CustomInspectors
 {
-    [CustomEditor(typeof(SpriteAtlasBundleLoader))]
-    public class SpriteBundleLoaderEditor : BundleLoaderEditor
+    public abstract class BaseSpriteBundleLoaderEditor<T> : BundleLoaderEditor where T: SpriteAtlasBundleLoader
     {
         private Sprite currentSprite;
         private Sprite initSprite;
@@ -25,7 +24,7 @@ namespace BundlesLoader.CustomInspectors
 
         private Sprite GetCurrentSprite()
         {
-            var obj = currentPath.serializedObject.targetObject as SpriteAtlasBundleLoader;
+            var obj = currentPath.serializedObject.targetObject as T;
             if (obj != null)
             {
                 var img = obj.GetComponent<Image>();
@@ -104,23 +103,19 @@ namespace BundlesLoader.CustomInspectors
 
         private void SetSprite(Sprite value)
         {
-            var obj = currentPath.serializedObject.targetObject as SpriteAtlasBundleLoader;
+            var obj = currentPath.serializedObject.targetObject as T;
             if(obj != null)
             {
-                var img = obj.GetComponent<Image>();
-                if (img != null)
-                {
-                    img.sprite = value;
-                }
+                obj.SetSprite(value);
             }
         }
 
         protected override string[] SetNames()
         {
-            var obj = currentPath.serializedObject.targetObject as SpriteAtlasBundleLoader;
+            var obj = currentPath.serializedObject.targetObject as T;
             if(obj == null)
             {
-                Debug.LogError($"No target object with specified type: {nameof(SpriteAtlasBundleLoader)}");
+                Debug.LogError($"No target object with specified type: {nameof(T)}");
                 return null;
             }