+++ Editor/CustomInspectors/ImageLoader/ImageSpriteBundleLoaderEditor.cs
@@ -1,0 +1,10 @@
+﻿using BundlesLoader.Bundles.Loaders.ImageLoader;
+using UnityEditor;
+
+namespace BundlesLoader.CustomInspectors.ImageLoader
+{
+    [CustomEditor(typeof(ImageSpriteAtlasBundleLoader))]
+    public class ImageSpriteBundleLoaderEditor : BaseSpriteBundleLoaderEditor<ImageSpriteAtlasBundleLoader>
+    {
+    }
+}