+++ Editor/CustomInspectors/SpriteRendererLoader/SpriteRendererSpriteBundleLoaderEditor.cs
@@ -1,0 +1,10 @@
+﻿using BundlesLoader.Bundles.Loaders.SpriteRendererLoader;
+using UnityEditor;
+
+namespace BundlesLoader.CustomInspectors.SpriteRendererLoader
+{
+    [CustomEditor(typeof(SpriteRendererSpriteAtlasBundleLoader))]
+    public class SpriteRendererSpriteBundleLoaderEditor : BaseSpriteBundleLoaderEditor<SpriteRendererSpriteAtlasBundleLoader>
+    {
+    }
+}