+++ Editor/CustomInspectors/SpriteRendererLoader/SpriteRendererTextureBundleLoaderEditor.cs
@@ -1,0 +1,10 @@
+﻿using BundlesLoader.Bundles.Loaders.SpriteRendererLoader;
+using UnityEditor;
+
+namespace BundlesLoader.CustomInspectors.SpriteRendererLoader
+{
+    [CustomEditor(typeof(SpriteRendererTextureBundleLoader))]
+    public class SpriteRendererTextureBundleLoaderEditor : BaseTextureBundleLoaderEditor<SpriteRendererTextureBundleLoader>
+    {
+    }
+}