+++ Runtime/Bundles/Loaders/SpriteRendererLoader/SpriteRendererTextureBundleLoader.cs
@@ -1,0 +1,21 @@
+using System;
+using UnityEngine;
+
+namespace BundlesLoader.Bundles.Loaders.SpriteRendererLoader
+{
+    [RequireComponent(typeof(SpriteRenderer))]
+    public class SpriteRendererTextureBundleLoader : TextureBundleLoader
+    {
+        private SpriteRenderer spriteRenderer;
+
+        public override void SetSprite(Sprite sprite)
+        {
+            if (spriteRenderer == null)
+            {
+                spriteRenderer = GetComponent<SpriteRenderer>();
+            }
+
+            spriteRenderer.sprite = sprite;
+        }
+    }
+}