+++ CHANGELOG.md
@@ -4,3 +4,8 @@ All notable changes to this project will be documented in this file.
 ## [1.0.0] - 2020-06-23
 ### Added
 - Added bundles loaders
+## [1.0.1] - 2020-07-08
+### Added
+- Added bundle loaders for SpriteRenderer component
+### Refactored
+- moved bundle loading logic to base classes